Combine common CSRF logic to a trait
